bootloader led blink command
