+++ src/bootloader.h
@@ -46,6 +46,7 @@
 typedef enum BOOT_OP_CODES {
     OP_WRITE = 'w',/* 0x77 */
     OP_ERASE = 'e',/* 0x65 */
+    OP_BLINK = 'b',/* 0x62 */
     OP_VERIFY = 's',/* 0x73 */
     OP_VERSION = 'v' /* 0x76 */
 } BOOT_OP_CODES;
@@ -64,8 +65,7 @@ typedef enum BOOT_STATUS {
 } BOOT_STATUS;
 
 
-uint8_t bootloader_firmware_verified(void);
-uint8_t bootloader_unlocked(void);
+void bootloader_jump(void);
 char *commander(const char *command);
